requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
rank-bm25>=0.2.2
networkx>=3.1
spacy>=3.6.0
tqdm>=4.65.0
zhipuai>=1.0.0